requirements.txt
fastapi==0.85.0
uvicorn[standard]
mlflow==2.2.2
scikit-learn==1.1.2
pandas==1.5.0
matplotlib==3.6.1
pytest==7.1.3
tensorflow<2.11
gitpython